refactor(ipc): split monolithic project.ts handler into 7 domain files
- Extract systemShell.ts (7 handlers: open-external, open-path, open-in-editor, check-command, check-directory, get-home-dir, get-tmp-dir)
- Extract editorConfig.ts (3 handlers: editor get/set config, discover)
- Extract agentCli.ts (8 handlers: CLI availability, agent versions, update settings, health check)
- Extract projectCrud.ts (19 handlers: project CRUD, switch, close, reopen, git init)
- Extract projectRecipes.ts (5 handlers: recipe CRUD)
- Extract terminalLayout.ts (8 handlers: terminals, sizes, tab groups, focus mode)
- Extract projectInRepoSettings.ts (4 handlers: CLAUDE.md read/write, in-repo settings toggle)
- Delete monolithic project.ts (1,524 lines) and update handlers.ts registry
- Update project.close.test.ts and handlers.registry.test.ts imports
