+++ electron/ipc/handlers.ts
@@ -3,7 +3,13 @@ import { registerWorktreeHandlers } from "./handlers/worktree.js";
 import { registerTerminalHandlers } from "./handlers/terminal.js";
 import { registerCopyTreeHandlers } from "./handlers/copyTree.js";
 import { registerAiHandlers } from "./handlers/ai.js";
-import { registerProjectHandlers } from "./handlers/project.js";
+import { registerSystemShellHandlers } from "./handlers/systemShell.js";
+import { registerEditorConfigHandlers } from "./handlers/editorConfig.js";
+import { registerAgentCliHandlers } from "./handlers/agentCli.js";
+import { registerProjectCrudHandlers } from "./handlers/projectCrud.js";
+import { registerProjectRecipesHandlers } from "./handlers/projectRecipes.js";
+import { registerTerminalLayoutHandlers } from "./handlers/terminalLayout.js";
+import { registerProjectInRepoSettingsHandlers } from "./handlers/projectInRepoSettings.js";
 import { registerGithubHandlers } from "./handlers/github.js";
 import { registerAppHandlers } from "./handlers/app.js";
 import { registerSidecarHandlers } from "./handlers/sidecar.js";
@@ -69,7 +75,13 @@ export function registerIpcHandlers(deps: HandlerDependencies): () => void {
     register(() => registerCopyTreeHandlers(deps));
     register(() => registerAiHandlers(deps));
     register(() => registerSlashCommandHandlers());
-    register(() => registerProjectHandlers(deps));
+    register(() => registerSystemShellHandlers(deps));
+    register(() => registerEditorConfigHandlers(deps));
+    register(() => registerAgentCliHandlers(deps));
+    register(() => registerProjectCrudHandlers(deps));
+    register(() => registerProjectRecipesHandlers(deps));
+    register(() => registerTerminalLayoutHandlers(deps));
+    register(() => registerProjectInRepoSettingsHandlers(deps));
     register(() => registerGithubHandlers(deps));
     register(() => registerAppHandlers(deps));
     register(() => registerSidecarHandlers(deps));
